Started to design gallery ppage

+++ scripts/publish_image.php
@@ -22,35 +22,45 @@ if ( isset( $_GET['path'] ) && !empty( $_GET['path'] ) )
 {
     $conn = pdo_connect();
     session_start();
-    if ( $conn !== NULL )
+
+    if ( $_SESSION['logged_in'] === TRUE )
     {
-        $login = $_SESSION['user_name'];
-        $date = date("Y-m-d H:i:s");
-        $query = $conn->prepare("INSERT INTO __camagru_posts.publications
-                                (
-                                    author,
-                                    publication_date
-                                )
-                                VALUES
-                                (
-                                    :author,
-                                    :publication_date
-                                )");
-        $query->bindParam(':author', $login);
-        $query->bindParam(':publication_date', $date);
-        $query->execute();
+        if ( $conn !== NULL )
+        {
+            $login = $_SESSION['user_name'];
+            $date = date("Y-m-d H:i:s");
+            $message = $_GET['message'];
+            $query = $conn->prepare("INSERT INTO __camagru_posts.publications
+                                    (
+                                        author,
+                                        publication_date,
+                                        message
+                                    )
+                                    VALUES
+                                    (
+                                        :author,
+                                        :publication_date,
+                                        :message
+                                    )");
+            $query->bindParam(':author', $login);
+            $query->bindParam(':publication_date', $date);
+            $query->bindParam(':message', $message);
+            $query->execute();
 
-        $other = $conn->prepare('SELECT id FROM __camagru_posts.publications
-                                WHERE
-                                (
-                                    author="' . $login. '" AND publication_date="' . $date . '")');
-        $other->execute();
-        $id = $other->fetch()['id'];
+            $other = $conn->prepare('SELECT id FROM __camagru_posts.publications
+                                    WHERE
+                                    (
+                                        author="' . $login. '" AND publication_date="' . $date . '")');
+            $other->execute();
+            $id = $other->fetch()['id'];
 
-        copy( $_GET['path'], "../posts/" . $id . ".png");
-        unlink( $_GET['pathl'] );
-        echo "OK";
+            copy( $_GET['path'], "../posts/" . $id . ".png");
+            unlink( $_GET['pathl'] );
+            echo "OK";
+        }
     }
+    else
+        echo "You must be logged in to publish creations";
     $conn = NULL;
 }
 else
